backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
anthropic==0.40.0
deepgram-sdk==3.5.0
python-socketio==5.11.0
python-engineio==4.8.0
eventlet==0.33.3
python-dotenv==1.0.0
